requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
psycopg2==2.7.7
python-dateutil==2.8.1
python-dotenv==0.10.3
python-http-client==3.2.6
pytz==2019.3
requests==2.23.0
sendgrid==6.2.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
